[SNAP-1760] changes to remove separate serialization header
added separate "channelSize()" method to SerializedDiskBuffer to distinguish between
data size and serialized size on disk (with serialization header)
