+++ gemfire-shared/src/main/java/com/gemstone/gemfire/internal/shared/ChannelBufferOutputStream.java
@@ -19,6 +19,7 @@ package com.gemstone.gemfire.internal.shared;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 import java.nio.channels.WritableByteChannel;
 import javax.annotation.Nonnull;
 
@@ -48,7 +49,11 @@ public class ChannelBufferOutputStream extends OutputStreamChannel {
   public ChannelBufferOutputStream(WritableByteChannel channel, int bufferSize)
       throws IOException {
     super(channel);
-    this.buffer = allocateBuffer(bufferSize);
+    if (bufferSize < 32) {
+      throw new IllegalArgumentException("buffer size " + bufferSize +
+          " should be at least 32");
+    }
+    this.buffer = allocateBuffer(bufferSize).order(ByteOrder.BIG_ENDIAN);
   }
 
   protected ByteBuffer allocateBuffer(int bufferSize) {
@@ -101,6 +106,18 @@ public class ChannelBufferOutputStream extends OutputStreamChannel {
   @Override
   public final int write(ByteBuffer src) throws IOException {
     return super.writeBuffered(src, this.buffer);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public final void writeInt(int v) throws IOException {
+    if (this.buffer.remaining() < 4) {
+      flushBufferBlocking(this.buffer);
+    }
+    // ByteBuffer will always be big-endian
+    this.buffer.putInt(v);
   }
 
   /**
